Разбил на модули код, согласно домашнему заданию и подключил в точке входа main.js. Указал тип файла main.js в index.html. Создал пустые модули, которые на мой взгляд будут согласно ТЗ.
